Split ir nodes into multiple translation units
